+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [#63](https://github.com/tboox/xmake/pull/63): More safer build and install xmake
 * [#61](https://github.com/tboox/xmake/pull/61): Check run command as root
 * Improve check toolchains and implement delay checking
+* Add user tips when scanning and generating `xmake.lua` automatically
 
 ## Bugs fixed
 
@@ -252,6 +253,7 @@
 * [#63](https://github.com/tboox/xmake/pull/63): 改进安装脚本，实现更加安全的构建和安装xmake
 * [#61](https://github.com/tboox/xmake/pull/61): 禁止在root权限下运行xmake命令，增强安全性
 * 改进工具链检测，通过延迟延迟检测提升整体检测效率
+* 当自动扫面生成`xmake.lua`时，添加更友好的用户提示，避免用户无操作
 
 ### Bugs修复